Remove mypy from pre-commit
Comment: Ok, running mypy in pre-commit in a useful and correct way
seems like a real hassle. By default, mypy will run in its own
pre-commit virtual environment, and so will not have access to
the dependencies of deltakit (which defeats 80% of its purpose
already). Now add the fact that it will ignore the excluded files
because modified files in the current commit are given explicitly to
mypy and I think we lost most of its usefulness. My opinion is that we
should remove mypy checks here, and let the CI perform a real
check (sadly)